[n_upsable][browser][browsershell] The VKB overlaps and covers the input field
:Release Notes:
The VKB overlaps and covers the input field in the system pages

:Detailed Notes:
The VKB overlaps and covers the input field in the system pages, such
as: Settings page & Site Filtering page.
This patch to scroll the input to the top of VKB

:Testing Performed:
1. Launch Enact Browser application
2. Open Settings page
3. Open Site-filtering page and select Blocked mode
4. Click to the Add input
Observe that VKB appears below the input and does not overlap.

:QA Notes:
Test cases: QWTC-12192, QWTC-13321

:Issues Addressed:
[WRR-14340] [First Round] The VKB overlaps and covers the input field

Change-Id: I78132c19e853d985777fc296e1f3086316cd6e23
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/461747
Reviewed-by: AI Reviewer <ai_reviewer@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Jeonghee27 Ahn <jeonghee27.ahn@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>